container- component 분리
