+++ front/gui/src/containers/ListContainer.js
@@ -1,12 +1,19 @@
-import React from 'react';
-import TodoTemplate from '../components/todo/TodoTemplate';
+import React, { useEffect } from 'react';
+import TodoList from '../components/todo/TodoList';
+import { useSelector, useDispatch } from 'react-redux';
+import { listTodos } from '../modules/list';
 
-const ListContainer = ({ date, handleDate }) => {
-  return (
-    <TodoTemplate date={date} changeDate={handleDate}>
-      <TodoTemplate></TodoTemplate>
-    </TodoTemplate>
-  );
+const ListContainer = () => {
+  const dispatch = useDispatch();
+  const { date } = useSelector((state) => ({
+    date: state.date.date,
+  }));
+
+  useEffect(() => {
+    dispatch(listTodos(date));
+  }, [date]);
+
+  return <TodoList></TodoList>;
 };
 
 export default ListContainer;
