+++ front/gui/src/index.js
@@ -18,7 +18,6 @@ const store = createStore(
 
 sagaMiddleware.run(rootSaga);
 
-// console.log(store.getState());
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
