+++ front/gui/src/lib/styles/TodoList.scss
@@ -1,5 +1,34 @@
-.TodoList {
-  min-height: 320px;
-  max-height: 513px;
-  overflow-y: auto;
+.todo-template {
+  width: 49%;
+  float: right;
+  /*flex로 내용 중앙 정렬 */
+  display: inline;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  @media screen and (max-width: 768px) {
+    /* 모바일에 사용될 스트일 시트를 여기에 작성합니다. */
+    float: none;
+  }
+
+  .app-title {
+    background: #ffd43b;
+    width: 100%;
+    height: 3rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    h3 {
+      margin: 1rem;
+      color: #495057;
+      margin-bottom: 1rem;
+    }
+  }
+
+  .TodoList {
+    min-height: 320px;
+    max-height: 513px;
+    overflow-y: auto;
+  }
 }
