+++ front/gui/src/lib/styles/TodoTemplate.scss
@@ -28,123 +28,5 @@
       font-size: 1.5rem;
       letter-spacing: 2px;
     }
-
-    .todo-template {
-      width: 49%;
-      float: right;
-      /*flex로 내용 중앙 정렬 */
-      display: inline;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-
-      @media screen and (max-width: 768px) {
-        /* 모바일에 사용될 스트일 시트를 여기에 작성합니다. */
-        float: none;
-      }
-
-      .app-title {
-        background: #ffd43b;
-        width: 100%;
-        height: 3rem;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        h3 {
-          margin: 1rem;
-          color: #495057;
-          margin-bottom: 1rem;
-        }
-      }
-      .content {
-        background: white;
-        width: 100%;
-      }
-    }
-
-    .calendar-area {
-      width: 49%;
-      float: left;
-      display: inline;
-      text-align: left;
-      @media screen and (max-width: 768px) {
-        /* 모바일에 사용될 스트일 시트를 여기에 작성합니다. */
-        float: none;
-        padding: 1rem;
-      }
-
-      .date-setting {
-        display: table;
-        padding: 0.5rem;
-        float: right;
-        button {
-          display: table-cell;
-          vertical-align: middle;
-          border: 0rem;
-          outline: 0rem;
-          background: white;
-          font-size: 1.5rem;
-          cursor: pointer;
-          &:hover {
-            color: #ffd43b;
-          }
-        }
-        span {
-          display: table-cell;
-          vertical-align: middle;
-          font-size: 1rem;
-        }
-      }
-      table {
-        width: 100%;
-        height: 500px;
-        border: 3px solid #fff3bf;
-        border-collapse: collapse;
-
-        td {
-          border: 3.5px solid #fff3bf;
-        }
-        .week {
-          text-align: center;
-          font-size: large;
-          font-weight: bold;
-          color: #495057;
-          padding-top: 1rem;
-          height: 1rem;
-        }
-
-        .day {
-          vertical-align: top;
-
-          width: 100px;
-          padding: 5px;
-          cursor: pointer;
-
-          &:nth-child(1) {
-            color: #f03e3e;
-          }
-          &:nth-child(7) {
-            color: #1c7ed6;
-          }
-          &:hover {
-            color: #fd7e14;
-          }
-        }
-      }
-      .title {
-        color: #495057;
-        background: #ffe066;
-        width: 100%;
-        height: 3rem;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        h3 {
-          margin: 1rem;
-          color: #495057;
-          margin-bottom: 1rem;
-        }
-      }
-    }
   }
 }
